+++ threedi_schematisation_editor/data_models.py
@@ -1174,7 +1174,11 @@ TABLE_MANNING = MappingProxyType(
     }
 )
 
-TABLE_SHAPES = [CrossSectionShape.TABULATED_RECTANGLE.value, CrossSectionShape.TABULATED_TRAPEZIUM.value]
+TABLE_SHAPES = [
+    CrossSectionShape.TABULATED_RECTANGLE.value,
+    CrossSectionShape.TABULATED_TRAPEZIUM.value,
+    CrossSectionShape.YZ.value,
+]
 
 MODEL_DEPENDENCIES = MappingProxyType(
     {