+++ threedi_schematisation_editor/enumerators.py
@@ -74,6 +74,8 @@ class CrossSectionShape(Enum):
     EGG = 3
     TABULATED_RECTANGLE = 5
     TABULATED_TRAPEZIUM = 6
+    YZ = 7
+    INVERTED_EGG = 8
 
 
 class ManholeShape(Enum):